requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langchain-openai==0.0.2
langsmith==0.0.92
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.57.4
packaging==23.2
pandas==2.2.3
Pillow==9.5.0
propcache==0.2.1
protobuf==4.25.5
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pyodbc==4.0.39
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.23
streamlit==1.24.0
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
yarl==1.18.3
zipp==3.21.0
